integration avec module user

+++ pidev-backend/src/main/java/tn/esprit/pidev/controller/evaluation/ScoreController.java
@@ -26,7 +26,7 @@ public class ScoreController {
     }
 
     @PostMapping("/addScore/{numQuiz}/{numUser}")
-    public Score addScore(@PathVariable("numQuiz")Long numQuiz,@PathVariable("numUser")Long numUser){
+    public Score addScore(@PathVariable("numQuiz")Long numQuiz,@PathVariable("numUser")Integer numUser){
         return IgScore.calculateScore(numQuiz,numUser);
     }
