+++ pidev-backend/src/main/java/tn/esprit/pidev/services/chat/IGestionDiscussion.java
@@ -5,19 +5,19 @@ import tn.esprit.pidev.entities.chat.Discussion;
 import java.util.List;
 
 public interface IGestionDiscussion {
-    public Discussion startDiscussionDuo(Long userStart, Long userEnd);
+    public Discussion startDiscussionDuo(Integer userStart, Integer userEnd);
 
-    public Discussion startDiscussionGroup(Long userStart, String title, List<Long> userList);
+    public Discussion startDiscussionGroup(Integer userStart, String title, List<Integer> userList);
 
-    public Discussion startDiscussionCommunity(Long userStart, String title, List<Long> userList, String discussionList);
+    public Discussion startDiscussionCommunity(Integer userStart, String title, List<Integer> userList, String discussionList);
 
-    public Discussion addUserToDiscussion(Long id, List<Long> userList);
+    public Discussion addUserToDiscussion(Long id, List<Integer> userList);
 
     public Discussion addDiscussionToCommunity(Long id, String discussionList);
 
-    public List<Discussion> retrieveAllDiscussions(Long id);
+    public List<Discussion> retrieveAllDiscussions(Integer id);
 
-    public List<Discussion> retrieveAllCommunities(Long id);
+    public List<Discussion> retrieveAllCommunities(Integer id);
 
     public Discussion renameDiscussion(Long id,String title);
